+++ dinero/tools/calculate_compound_interest.py
@@ -1,5 +1,6 @@
 from dinero import Dinero
-from dinero.exceptions import InvalidOperationError
+
+from ._validators import ToolValidators
 
 
 def calculate_compound_interest(
@@ -30,26 +31,13 @@ def calculate_compound_interest(
         >>> calculate_compound_interest(principal, interest_rate, duration, compound_frequency)
         Dinero(648.34)
     """
-    validate_inputs(principal, interest_rate, duration, compound_frequency)
+    validate = ToolValidators()
+    validate.compound_interest_inputs(
+        principal, interest_rate, duration, compound_frequency
+    )
 
     # Calculate the total interest using the formula: A = P * (1 + r/n)^(n*t)
     n = compound_frequency
     r = interest_rate / 100
     t = duration
     return principal * ((1 + r / n) ** (n * t) - 1)
-
-
-def validate_inputs(
-    principal: Dinero, interest_rate: float, duration: int, compound_frequency: int
-) -> None:
-    if not isinstance(principal, Dinero):
-        raise InvalidOperationError(InvalidOperationError.operation_msg)
-
-    if not isinstance(interest_rate, (float, int)) or interest_rate <= 0:
-        raise ValueError("Interest rate must be a positive float.")
-
-    if not isinstance(duration, int) or duration <= 0:
-        raise ValueError("Duration must be a positive integer.")
-
-    if not isinstance(compound_frequency, int) or compound_frequency <= 0:
-        raise ValueError("Compound frequency must be a positive integer.")