+++ dinero/tools/calculate_vat.py
@@ -1,5 +1,6 @@
 from dinero import Dinero
-from dinero.exceptions import InvalidOperationError
+
+from ._validators import ToolValidators
 
 
 def calculate_vat(amount: Dinero, vat_rate: int | float) -> Dinero:
@@ -24,19 +25,9 @@ def calculate_vat(amount: Dinero, vat_rate: int | float) -> Dinero:
         >>> vat_amount.format(symbol=True, currency=True)
         '$6.76 USD'
     """
-    validate_inputs(amount, vat_rate)
+    validate = ToolValidators()
+    validate.vat_inputs(amount, vat_rate)
     divisor = 1 + (vat_rate / 100)
     amount_without_vat = amount / divisor
     vat_amount = amount - amount_without_vat
     return vat_amount
-
-
-def validate_inputs(amount: Dinero, vat_rate: int | float) -> None:
-    if not isinstance(amount, Dinero):
-        raise InvalidOperationError(InvalidOperationError.operation_msg)
-
-    if not isinstance(vat_rate, (int, float)):
-        raise TypeError("The vat_rate argument must be a number.")
-
-    if vat_rate < 0:
-        raise ValueError("The vat_rate argument cannot be negative.")